newly pushed

+++ database/migrations/2025_02_16_155458_create_customers_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
             $table->text('customer_code');            
             $table->text('customer_name');
             $table->text('customer_singkatan');
-            $table->integer('customer_phone')->length(13);
+            $table->bigInteger('customer_phone')->length(13);
             $table->string('customer_email')->unique();
             $table->text('customer_address');
-            $table->integer('customer_npwp')->length(16);
+            $table->bigInteger('customer_npwp')->length(16);
             $table->string('customer_contact');            
             $table->timestamps();
         });
